Reading the String and Checking for Palindrome
The main concept is to read the string after converting it to the lowercase and removing all the Non-Alphanumeric characters, by its each character from its starting character at the 0th position and comparing it to the last character of the string at the position of string.length-1. Continuing this until the half the length of string.